+++ Sources/KeyboardKit/Actions/DeleteBackwardRange.swift
@@ -9,8 +9,8 @@
 import Foundation
 
 /**
- This enum can be used to vary how the backspace action will
- behave when pressing and holding the backspace key.
+ This enum can be used to control how much the backspace key
+ will delete when it repeats its action.
  */
 public enum DeleteBackwardRange {