+++ Sources/KeyboardKit/Autocomplete/AutocompleteToolbarItemTitle.swift
@@ -87,7 +87,7 @@ struct AutocompleteToolbarItemTitle_Previews: PreviewProvider {
     
     static let previewSuggestions: [AutocompleteSuggestion] = [
         AutocompleteSuggestion(text: "Foo", isUnknown: true),
-        AutocompleteSuggestion(text: "Bar", isAutocomplete: true),
+        AutocompleteSuggestion(text: "Bar", isAutocorrect: true),
         AutocompleteSuggestion(text: "", title: "Baz", subtitle: "Recommended")]
 }